Add LICENSE, READM.md, and update cabal and packaging
NOTE: Older commits have a placeholder auto-generated `LICENSE` file that uses BSD 3-clause. These versions were never distributed prior to this commit, which adds the proper licenses. As such, please disregard prior commits that are licensed improperly, not just because they're buggy and incorrect, but because they are retroactively licensed the same as this commit.