Fix non-termination bug mentioned in previous commit
The issue was:

```haskell
typingCheck ctx r (SrcLambda (argName:otherArgNames) lambdaBody) (PI piArgName piArgType piReturnType) = -- FunI-2
    typingCheck ctx r (SrcLambda [argName] (SrcLambda otherArgNames lambdaBody)) (PI piArgName piArgType piReturnType)
```

instead of:

```haskell
typingCheck ctx r (SrcLambda (argName:otherArgNames) lambdaBody) typeVal = -- FunI-2
    typingCheck ctx r (SrcLambda [argName] (SrcLambda otherArgNames lambdaBody)) typeVal
```

This would cause infinite loops if a non-Pi type was provided or if a Pi type with *less* than the correct number of args was provided as the type for a lambda expression.

This is because it would not ever choose `FunI-1` as it was desugaring and then eventually ended up where `FunI-2` was calling itself with `lambdaBody` equaling `SrcLambda` and the arg list just being `argName` (with `otherArgNames` empty).
The issue was because it would disqualify `FunI-1`, it would never choose the `[argName]` overload of the lambda over the `(argName:otherArgNames)` overload.