Redo module layout, refactor multiple functions, and revamp tests
This commit is a big clean-up commit. It optimizes and reworks the module layouts, exports, and imports so that it is much cleaner and more consistent. Previously, the placement of functions in different files was quite random and messy.
Additionally, multiple functions have been refactored to be shorter or use more idiomatic styles (such as removing custom `lookup` functions).