Add error messages and update cabal & licensing
All files now have a copyright header. The README.md was also updated to be more clear.

The cabal file has been updated so that `cabal test` now works for any Haskell tests that work (I fixed the simple parsing tests). This involved splitting the public and private interface.

The biggest change is that I have now replaced all uses of `Maybe a` with `Either Error a`. The type checker, etc. now reports actual error messages instead of just `Nothing`.

There are improvements that could be made such as better error messages in the parser, multi-error reporting for a single file, line/column reporting in type checking error messages, etc.; but I have decided that given the simple nature of this language, the current error messages are sufficient. Unless any bugs or other issues surface, this should be the last commit of this project (notwithstanding if I fork it and extend its type system to MLTT later).